refactor: deduplicate operator parsing in quspin-py
Extract the shared `parse_terms`, `max_site_from_terms`, `extract_coeff`,
and `Term`/`Terms` type aliases from pauli.rs into operator/mod.rs. The
new `parse_terms_generic` function is parameterized over the op type
(via `ParseOp` trait) and entry constructor (via closure), eliminating
~150 lines of near-identical code across pauli.rs, boson.rs, and
fermion.rs.

Closes #29
