+++ crates/quspin-py/src/operator/pauli.rs
@@ -1,15 +1,11 @@
 use crate::error::Error;
-use crate::operator::{as_c64_vec, with_space_inner, with_two_space_inners, write_c64_back};
-use num_complex::Complex;
+use crate::operator::{
+    Terms, as_c64_vec, max_site_from_terms, parse_terms_generic, with_space_inner,
+    with_two_space_inners, write_c64_back,
+};
 use numpy::{Complex64, PyArray1};
 use pyo3::prelude::*;
 use quspin_core::operator::pauli::{HardcoreOp, HardcoreOperator, HardcoreOperatorInner, OpEntry};
-use smallvec::SmallVec;
-
-/// A single term: a list of `(op_str, bonds)` pairs sharing one cindex.
-pub(crate) type Term = Vec<(String, Vec<Vec<PyObject>>)>;
-/// All terms passed to a constructor (one per cindex).
-pub(crate) type Terms = Vec<Term>;
 
 /// Python-facing Pauli / hardcore-boson / spin-½ operator.
 ///
@@ -46,12 +42,14 @@ impl PyPauliOperator {
         let use_u8 = max_cindex <= 255 && max_site <= 255;
 
         if use_u8 {
-            let entries = parse_terms::<u8>(py, &terms).map_err(Error::from)?;
+            let entries = parse_terms_generic::<u8, HardcoreOp, _, _>(py, &terms, OpEntry::new)
+                .map_err(Error::from)?;
             Ok(PyPauliOperator {
                 inner: HardcoreOperatorInner::Ham8(HardcoreOperator::new(entries)),
             })
         } else if max_cindex <= 65535 && max_site <= 65535 {
-            let entries = parse_terms::<u16>(py, &terms).map_err(Error::from)?;
+            let entries = parse_terms_generic::<u16, HardcoreOp, _, _>(py, &terms, OpEntry::new)
+                .map_err(Error::from)?;
             Ok(PyPauliOperator {
                 inner: HardcoreOperatorInner::Ham16(HardcoreOperator::new(entries)),
             })
@@ -141,92 +139,4 @@ impl PyPauliOperator {
             self.inner.num_cindices(),
         )
     }
-}
-
-// ---------------------------------------------------------------------------
-// Shared helpers
-// ---------------------------------------------------------------------------
-
-/// Extract the max site index from all bonds across all terms.
-pub(crate) fn max_site_from_terms(py: Python<'_>, terms: &[Term]) -> PyResult<usize> {
-    let mut max = 0usize;
-    for term in terms {
-        for (_, bonds) in term {
-            for bond in bonds {
-                // bond = [coeff, site0, site1, ...]  — skip index 0 (coeff)
-                for obj in bond.iter().skip(1) {
-                    let site: u32 = obj.bind(py).extract()?;
-                    max = max.max(site as usize);
-                }
-            }
-        }
-    }
-    Ok(max)
-}
-
-/// Extract a complex coefficient from a Python scalar (int, float, or complex).
-pub(crate) fn extract_coeff(py: Python<'_>, obj: &PyObject) -> PyResult<Complex<f64>> {
-    let bound = obj.bind(py);
-    if let Ok(z) = bound.extract::<Complex<f64>>() {
-        return Ok(z);
-    }
-    let re: f64 = bound.extract()?;
-    Ok(Complex::new(re, 0.0))
-}
-
-fn parse_terms<C: Copy + Ord + TryFrom<usize>>(
-    py: Python<'_>,
-    terms: &[Term],
-) -> Result<Vec<OpEntry<C>>, quspin_core::error::QuSpinError>
-where
-    <C as TryFrom<usize>>::Error: std::fmt::Debug,
-{
-    let mut entries = Vec::new();
-    for (cindex_usize, term) in terms.iter().enumerate() {
-        let cindex = C::try_from(cindex_usize).map_err(|_| {
-            quspin_core::error::QuSpinError::ValueError(format!(
-                "cindex {cindex_usize} out of range for chosen index type"
-            ))
-        })?;
-        for (op_str, bonds) in term {
-            for bond in bonds {
-                if bond.is_empty() {
-                    return Err(quspin_core::error::QuSpinError::ValueError(
-                        "each bond must be [coeff, site0, site1, ...]".to_string(),
-                    ));
-                }
-                let coeff = extract_coeff(py, &bond[0]).map_err(|e| {
-                    quspin_core::error::QuSpinError::ValueError(format!("bond coefficient: {e}"))
-                })?;
-                let sites: Vec<u32> = bond[1..]
-                    .iter()
-                    .map(|s| {
-                        s.bind(py).extract::<u32>().map_err(|e| {
-                            quspin_core::error::QuSpinError::ValueError(format!(
-                                "bond site index: {e}"
-                            ))
-                        })
-                    })
-                    .collect::<Result<_, _>>()?;
-                if op_str.len() != sites.len() {
-                    return Err(quspin_core::error::QuSpinError::ValueError(format!(
-                        "op_str length {} != number of sites {}",
-                        op_str.len(),
-                        sites.len()
-                    )));
-                }
-                let mut ops: SmallVec<[(HardcoreOp, u32); 4]> = SmallVec::new();
-                for (ch, &site) in op_str.chars().zip(sites.iter()) {
-                    let op = HardcoreOp::from_char(ch).ok_or_else(|| {
-                        quspin_core::error::QuSpinError::ValueError(format!(
-                            "unknown operator character '{ch}'; expected one of x, y, z, +, -, n"
-                        ))
-                    })?;
-                    ops.push((op, site));
-                }
-                entries.push(OpEntry::new(cindex, coeff, ops));
-            }
-        }
-    }
-    Ok(entries)
 }
